+++ packages/react/src/slider/root/SliderRootContext.ts
@@ -1,12 +1,83 @@
 'use client';
 import * as React from 'react';
-import type { SliderRoot } from './SliderRoot';
-import type { useSliderRoot } from './useSliderRoot';
+import type { Orientation } from '../../utils/types';
+import { type CompositeMetadata } from '../../composite/list/CompositeList';
+import type { ThumbMetadata } from '../thumb/SliderThumb';
+import type { SliderRoot, FingerPosition, FingerState } from './SliderRoot';
 
-export interface SliderRootContext extends Omit<useSliderRoot.ReturnValue, 'getRootProps'> {
+export interface SliderRootContext {
+  /**
+   * The index of the active thumb.
+   */
+  active: number;
+  /**
+   * Function to be called when drag ends. Invokes onValueCommitted.
+   */
+  commitValue: (newValue: number | readonly number[], event: Event) => void;
+  dragging: boolean;
+  disabled: boolean;
   format?: Intl.NumberFormatOptions;
+  getFingerState: (
+    fingerPosition: FingerPosition | null,
+    shouldCaptureThumbIndex?: boolean,
+    offset?: number,
+  ) => FingerState | null;
+  handleInputChange: (
+    valueInput: number,
+    index: number,
+    event: React.KeyboardEvent | React.ChangeEvent,
+  ) => void;
+  labelId?: string;
+  /**
+   * The large step value of the slider when incrementing or decrementing while the shift key is held,
+   * or when using Page-Up or Page-Down keys. Snaps to multiples of this value.
+   * @default 10
+   */
+  largeStep: number;
+  lastChangedValueRef: React.RefObject<number | readonly number[] | null>;
+  /**
+   * The maximum allowed value of the slider.
+   */
+  max: number;
+  /**
+   * The minimum allowed value of the slider.
+   */
+  min: number;
+  /**
+   * The minimum steps between values in a range slider.
+   */
+  minStepsBetweenValues: number;
+  name: string;
+  /**
+   * The component orientation.
+   * @default 'horizontal'
+   */
+  orientation: Orientation;
+  registerSliderControl: (element: HTMLElement | null) => void;
+  setActive: React.Dispatch<React.SetStateAction<number>>;
+  setDragging: React.Dispatch<React.SetStateAction<boolean>>;
+  setThumbMap: React.Dispatch<
+    React.SetStateAction<Map<Node, CompositeMetadata<ThumbMetadata> | null>>
+  >;
+
+  /**
+   * Callback fired when dragging and invokes onValueChange.
+   */
+  setValue: (newValue: number | number[], activeThumb: number, event: Event) => void;
   state: SliderRoot.State;
+  /**
+   * The step increment of the slider when incrementing or decrementing. It will snap
+   * to multiples of this value. Decimal values are supported.
+   * @default 1
+   */
+  step: number;
   tabIndex: number | null;
+  thumbMap: Map<Node, CompositeMetadata<ThumbMetadata> | null>;
+  thumbRefs: React.RefObject<(HTMLElement | null)[]>;
+  /**
+   * The value(s) of the slider
+   */
+  values: readonly number[];
 }
 
 export const SliderRootContext = React.createContext<SliderRootContext | undefined>(undefined);
